Client component to server component
