+++ src/features/Post/index.tsx
@@ -1,10 +1,6 @@
-'use client';
-
 import {Post_type} from '@/features/Post/types/Post_type';
-import {Calendar, ArrowUp, ArrowDown, MessageCircle} from 'lucide-react';
 import {Avatar, AvatarFallback, AvatarImage} from '@/components/ui/avatar';
-import {useState} from 'react';
-import ReplyForm from './components/ReplyForm';
+import FooterRight from './components/FooterMechanicComponent'; 
 import {
   Card,
   CardHeader,
@@ -13,24 +9,11 @@ import {
   CardFooter,
 } from '@/components/ui/card';
 
-export default function Post({...post}: Post_type) {
-  const [showReply, setShowReply] = useState(false);
-
-  const handleMessage = () => {
-    setShowReply(prev => !prev);
-  };
-
-  const handleUpvote = () => {
-    console.log('Upvote clicked for post', post.id);
-  };
-
-  const handleDownvote = () => {
-    console.log('Downvote clicked for post', post.id);
-  };
+export default function Post({...post}: Post_type) { 
 
   return (
     <div>
-      <Card className='rounded-lg bg-white shadow-md'>
+      <Card className='rounded-lg bg-white shadow-md pb-0'>
         <CardHeader className='flex items-center justify-between gap-4 pb-4'>
           <div className='flex gap-4'>
             <Avatar>
@@ -49,39 +32,11 @@ export default function Post({...post}: Post_type) {
 
         <CardDescription className='px-4 py-2'>{post.content}</CardDescription>
 
-        <CardFooter className='flex items-center justify-between px-4 py-2 text-sm text-gray-500'>
-          <div className='flex items-center gap-2'>
-            <Calendar className='h-4 w-4' />
-            {post.createdAt && new Date(post.createdAt).toLocaleDateString()}
-          </div>
-
-          <div className='flex items-center gap-1'>
-            <button
-              onClick={handleMessage}
-              className='flex w-25 cursor-pointer items-center gap-2 hover:text-black'
-            >
-              <MessageCircle className='h-4 w-4' />
-              <span>Reply</span>
-            </button>
-
-            <button
-              onClick={handleUpvote}
-              className='cursor-pointer hover:text-green-500'
-            >
-              <ArrowUp className='h-4 w-4' />
-            </button>
-            <span>{post.rating ?? 0}</span>
-            <button
-              onClick={handleDownvote}
-              className='cursor-pointer hover:text-red-500'
-            >
-              <ArrowDown className='h-4 w-4' />
-            </button>
-          </div>
+        <CardFooter className='px-0'>  
+          <FooterRight post={post} />
         </CardFooter>
       </Card>
 
-      {showReply && <ReplyForm postId={post.id} />}
     </div>
   );
 }
